Moved original code to appropriate packages
