+++ app/src/main/java/com/lambdaschool/android/organization/util/DateUtils.kt
@@ -1,4 +1,4 @@
-package com.lambdaschool.android.organization
+package com.lambdaschool.android.organization.util
 
 import java.text.SimpleDateFormat
 import java.util.*
